requirements.txt
streamlit
PyPDF2
nltk
gensim
transformers
sentence-transformers
groq
faiss-cpu

